Handle deleting a copied container and cleanup
