--- panoramapublic/src/org/labkey/panoramapublic/SymlinkHandler.java
@@ -1,8 +1,0 @@
-package org.labkey.panoramapublic;
-
-import java.nio.file.Path;
-
-public interface SymlinkHandler
-{
-    void handleSymlink(Path link, Path target);
-}
